services/chimera_core/requirements.txt
fastapi
uvicorn
pyyaml
httpx
python-dotenv
pydantic-settings
websockets
